# Add --no-syntax option