@Enselic 2025-07-16
can we somehow be "smart" so users don't need a new option?