@Enselic 2025-08-16
for my understanding: why not use `cat` for that use case?