@MoSal 2025-08-16
This should be closed actually, together with the other PR. This one is just wrong. And I've come up with a new approach that works for #3101. I just want to explore alternative ways to do syntax highlighting first, because syntect is the only slow part remaining.